Move around email helper code to make MonitorHelper small for rubocop
