--- spec/support/monitor/gmail_helper.rb
@@ -1,18 +1,0 @@
-require 'gmail'
-
-class GmailHelper
-  attr_reader :gmail
-
-  def initialize(email, password)
-    @gmail = Gmail.connect!(email, password)
-  end
-
-  def inbox_unread
-    gmail.inbox.emails(:unread)
-  end
-
-  def inbox_clear
-    inbox_unread.each(&:read!)
-    gmail.inbox.emails(:read).each(&:delete!)
-  end
-end
